My name is Poppy, and I am 3.5 :) - Woof!

+++ 0x0E-structures_typedef/4-new_dog.c
@@ -9,7 +9,7 @@
  */
 dog_t *new_dog(char *name, float age, char *owner)
 {
-	size_t n, o, i;
+	unsigned int n, o, i;
 	dog_t *dog;
 
 	if (name == NULL || owner == NULL)
@@ -44,5 +44,5 @@ dog_t *new_dog(char *name, float age, char *owner)
 	for (i = 0; i < o; i++)
 		dog->owner[i] = owner[i];
 	return (dog);
-	
+
 }
